+++ examples/rainbowkit-tailwind/components/FocusSection/FocusSection.tsx
@@ -1,0 +1,10 @@
+const FocusSection = (props: any) => {
+    const {focus} = props
+    return (
+        <div>
+            {focus ? focus.map((item: string) => <p key={item}>{item}</p>) : ""}
+        </div>
+    )
+}
+
+export default FocusSection