+++ examples/rainbowkit-tailwind/components/FocusSection/index.ts
@@ -1,0 +1,3 @@
+import FocusSection from "./FocusSection";
+
+export default FocusSection