+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2022-11-25
+
+### functional mvp
+
+- `src/evaluators` - music platforms put their logic for `evaluate` to `compareJson` & `calculateScore`.
+- `src/schemas/` - music platforms put their schema examples there.
+- `lib/compareJson` - shared util to find the difference between 2 JSON files.
+  - TODO: handle nested objects.
+
 ## [0.0.1] - 2022-11-24
 
 ### Update README